+++ crates/parol_runtime/CHANGELOG.md
@@ -17,6 +17,10 @@ Be aware that this project is still v0.y.z which means that anything can change 
 We defined for this project that while being on major version zero we mark incompatible changes with
 new minor version numbers. Please note that this is no version handling covered by `Semver`.
 
+## 1.0.2 - Not released yet
+
+- Error recovery on generated LL(k) parsers can now be disabled
+
 ## 1.0.1 - 2024-10-08
 
 - Updated dependencies